* Refactor and remove redundant checks * Cover additional tests * Add more tests and improve coverage * Add comments and code annotations * Rearrange tests * Refactor code to change to folders * Rename Mock E2E tests * Add more tests for input bindings
