+++ Directory.Packages.props
@@ -1,5 +1,6 @@
 <Project>
   <ItemGroup>
+    <PackageVersion Include="coverlet.collector" Version="3.2.0" />
     <PackageVersion Include="Microsoft.AspNet.Mvc">
       <Version>5.2.9</Version>
     </PackageVersion>
@@ -9,9 +10,9 @@
     <PackageVersion Include="Microsoft.AspNetCore.Mvc.Core">
       <Version>2.2.5</Version>
     </PackageVersion>
-    <PackageVersion Include="Microsoft.Azure.Kusto.Data" Version="11.0.0" />
+    <PackageVersion Include="Microsoft.Azure.Kusto.Data" Version="11.1.0" />
     <PackageVersion Include="Microsoft.Azure.Kusto.Ingest">
-      <Version>11.0.0</Version>
+      <Version>11.1.0</Version>
     </PackageVersion>
     <PackageVersion Include="Microsoft.Azure.WebJobs.Core">
       <Version>3.0.33</Version>
@@ -33,9 +34,11 @@
     <PackageVersion Include="Microsoft.Data.SqlClient" Version="3.0.1" />
     <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.1.1" />
     <PackageVersion Include="morelinq" Version="3.3.2" />
+    <PackageVersion Include="System.Linq.Async" Version="6.0.1" />
+    <PackageVersion Include="System.Linq.Async.Queryable" Version="6.0.1" />
     <PackageVersion Include="System.Runtime.Caching" Version="4.7.0" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
-    <PackageVersion Include="Moq" Version="4.18.2" />
+    <PackageVersion Include="Moq" Version="4.18.3" />
     <PackageVersion Include="xunit" Version="2.4.0" />
     <PackageVersion Include="xunit.core">
       <Version>2.3.1</Version>
@@ -43,9 +46,9 @@
     <PackageVersion Include="xunit.extensibility.core">
       <Version>2.3.1</Version>
     </PackageVersion>
-    <PackageVersion Include="xunit.runner.visualstudio" Version="2.3.1" />
+    <PackageVersion Include="xunit.runner.visualstudio" Version="2.4.5" />
     <PackageVersion Include="Microsoft.Azure.WebJobs.Script.ExtensionsMetadataGenerator" Version="1.2.3" />
-    <PackageVersion Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageVersion Include="Newtonsoft.Json" Version="13.0.2" />
     <PackageVersion Include="Microsoft.AspNetCore.Http" Version="2.2.2" />
     <PackageVersion Include="BenchmarkDotNet" Version="0.13.1" />
   </ItemGroup>
