+++ src/Config/KustoBindingExtension.cs
