+++ src/Services/KustoBindingUtilities.cs
