+++ src/Services/KustoIngestContext.cs
@@ -1,15 +1,18 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using Kusto.Data.Common;
+using Kusto.Ingest;
 using Microsoft.Azure.WebJobs.Kusto;
 
 namespace Microsoft.Azure.WebJobs.Extensions.Kusto
 {
-    internal class KustoQueryContext
+    /// <summary>
+    /// Wraps around the attribute and the ingest client. Makes it easy to extract fields from KustoAttribute and ingest data provided in the attribute
+    /// </summary>
+    internal class KustoIngestContext
     {
         public KustoAttribute ResolvedAttribute { get; set; }
 
-        public ICslQueryProvider QueryProvider { get; set; }
+        public IKustoIngestClient IngestService { get; set; }
     }
 }
