+++ src/Services/KustoQueryContext.cs
@@ -1,15 +1,18 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using Kusto.Ingest;
+using Kusto.Data.Common;
 using Microsoft.Azure.WebJobs.Kusto;
 
 namespace Microsoft.Azure.WebJobs.Extensions.Kusto
 {
-    internal class KustoIngestContext
+    /// <summary>
+    /// Wraps around the attribute and the query client. Makes it easy to extract fields from KustoAttribute and execute it on the client
+    /// </summary>
+    internal class KustoQueryContext
     {
         public KustoAttribute ResolvedAttribute { get; set; }
 
-        public IKustoIngestClient IngestService { get; set; }
+        public ICslQueryProvider QueryProvider { get; set; }
     }
 }
