+++ test/Bindings/KustoAsyncCollectorTests.cs
