--- test/Resources/I1.json
@@ -1,36 +1,0 @@
-﻿{
-  "Tables": [
-    {
-      "TableName": "Table_0",
-      "Columns": [
-        {
-          "ColumnName": "ProductID",
-          "DataType": "Int32",
-          "ColumnType": "int"
-        },
-        {
-          "ColumnName": "Name",
-          "DataType": "String",
-          "ColumnType": "string"
-        },
-        {
-          "ColumnName": "Cost",
-          "DataType": "Int32",
-          "ColumnType": "int"
-        }
-      ],
-      "Rows": [
-        [
-          1,
-          "I1",
-          100
-        ],
-        [
-          2,
-          "I1",
-          200
-        ]
-      ]
-    }
-  ]
-}
